Add doc for external API. Refactor sm3 & sm4 API.
Supplementary documentation for external APIs.
Refactor sm3, sm4 API, add ctx_new(), ctx_free() API.
Rename some API with the prefix tsm_.

+++ include/tongsuo/oscore.h
@@ -19,12 +19,6 @@ extern "C" {
 # include <tongsuo/oscore_context.h>
 # include <tongsuo/mem.h>
 
-/**
- * @ingroup internal_api
- * @addtogroup oscore_internal
- * @{
- */
-
 # define TSM_OSCORE_DEFAULT_REPLAY_WINDOW 32
 
 /* Estimate your header size, especially when using Proxy-Uri. */
@@ -34,18 +28,16 @@ extern "C" {
 # define OSCORE_DECRYPTION_ERROR    100
 # define PACKET_SERIALIZATION_ERROR 102
 
-/*
- * tsm_oscore_encode_option_value
- */
+/* Encodes the OSCORE option value to option_buffer with length option_buf_len. Returns the written
+ * length. */
 size_t tsm_oscore_encode_option_value(uint8_t *option_buffer,
                                       size_t option_buf_len,
                                       cose_encrypt0_t *cose,
                                       uint8_t group,
                                       uint8_t appendix_b_2);
 
-/*
- * Decodes the OSCORE option value and places decoded values into the provided
- * cose structure */
+/* Decodes the OSCORE option value and places decoded values into the provided
+ * cose structure. Returns TSM_OK if successful, error code otherwise. */
 int tsm_oscore_decode_option_value(const uint8_t *option_value,
                                    size_t option_len,
                                    cose_encrypt0_t *cose);
@@ -56,7 +48,36 @@ size_t tsm_oscore_prepare_aad(const uint8_t *external_aad_buffer,
                               size_t external_aad_len,
                               uint8_t *aad_buffer,
                               size_t aad_size);
-
+/*
+ * Build the CBOR for external_aad
+ *
+ * external_aad = bstr .cbor aad_array
+ *
+ * No group mode
+ * aad_array = [
+ *   oscore_version : uint,
+ *   algorithms : [ alg_aead : int / tstr ],
+ *   request_kid : bstr,
+ *   request_piv : bstr,
+ *   options : bstr,
+ * ]
+ *
+ * Group mode
+ * aad_array = [
+ *   oscore_version : uint,
+ *   algorithms : [alg_aead : int / tstr / null,
+ *                 alg_signature_enc : int / tstr / null,
+ *                 alg_signature : int / tstr / null,
+ *                 alg_pairwise_key_agreement : int / tstr / null],
+ *   request_kid : bstr,
+ *   request_piv : bstr,
+ *   options : bstr,
+ *   request_kid_context : bstr,
+ *   OSCORE_option: bstr,
+ *   sender_public_key: bstr,        (initiator's key)
+ *   gm_public_key: bstr / null
+ * ]
+ */
 size_t tsm_oscore_prepare_e_aad(oscore_ctx_t *ctx,
                                 cose_encrypt0_t *cose,
                                 const uint8_t *oscore_option,
@@ -65,24 +86,23 @@ size_t tsm_oscore_prepare_e_aad(oscore_ctx_t *ctx,
                                 uint8_t *external_aad_ptr,
                                 size_t external_aad_size);
 
-/* Creates Nonce */
+/* Creates nonce and writes to buffer with length size. */
 void tsm_oscore_generate_nonce(cose_encrypt0_t *ptr,
                                oscore_ctx_t *ctx,
                                uint8_t *buffer,
                                uint8_t size);
 
-/*Return 1 if OK, Error code otherwise */
-uint8_t tsm_oscore_validate_sender_seq(oscore_recipient_ctx_t *ctx, cose_encrypt0_t *cose);
+/* Validate the sequence of sender. Returns TSM_OK if OK, error code otherwise. */
+int tsm_oscore_validate_sender_seq(oscore_recipient_ctx_t *ctx, cose_encrypt0_t *cose);
 
-/* Return 0 if SEQ MAX, return 1 if OK */
-uint8_t tsm_oscore_increment_sender_seq(oscore_ctx_t *ctx);
-
+/* Increases the sequence of sender. Returns TSM_OK on succes, error code on error. */
+int tsm_oscore_increment_sender_seq(oscore_ctx_t *ctx);
 /* Restore the sequence number and replay-window to the previous state. This is
  * to be used when decryption fail. */
 void tsm_oscore_roll_back_seq(oscore_recipient_ctx_t *ctx);
-
+/* Create a OSCORE conf object which should be freed by calling tsm_oscore_conf_free() after use. */
 TSM_OSCORE_CONF *tsm_oscore_conf_new(void);
-
+/* Frees up the OSCORE conf object. */
 void tsm_oscore_conf_free(TSM_OSCORE_CONF *oscf);
 
 # ifdef __cplusplus
